requirements.txt
streamlit==1.32.0
google-cloud-storage==2.13.0
langchain>=0.0.310
langchain-openai>=0.0.2
langchain-community>=0.0.17
langchain-text-splitters>=0.0.1
openai>=1.3.0
pdfminer.six==20221105
pymupdf==1.23.6
chromadb==0.4.18
tiktoken==0.5.2
pandas==2.1.3
pypdf==3.17.1
python-magic==0.4.27
unstructured==0.10.30
pdf2image==1.16.3
pytesseract==0.3.10